+++ tests/test_main.py
@@ -200,7 +200,7 @@ class TestParseArgs(unittest.TestCase):
         self.assertEqual(cm.exception.code, 2)
         self.assertTrue(
             "error: the following arguments are required: -c/--config, "
-            "-a/--application-id, -r/--run-id" in f.getvalue()
+            "-a/--application-id" in f.getvalue()
         )
 
     def test_parse_args_valid_arguments(self):
@@ -285,6 +285,7 @@ class TestParseArgs(unittest.TestCase):
                 "quotechar='\"', "
                 "run_description='', "
                 "run_id=123, "
+                "run_id_file=None, "
                 "skip_footer=0, "
                 "skip_header=0, "
                 "truncate=[], "
@@ -686,7 +687,7 @@ class TestInit(unittest.TestCase):
         self.assertEqual(cm.exception.code, 2)
         self.assertTrue(
             "error: the following arguments are required: -c/--config, "
-            "-a/--application-id, -r/--run-id" in f.getvalue()
+            "-a/--application-id" in f.getvalue()
         )
 
     def test_init_existing_metadata_file(self):
@@ -1017,9 +1018,7 @@ class TestInit(unittest.TestCase):
         self.assertEqual(cm1.exception.code, 223)
         self.assertEqual(
             cm2.output,
-            [
-                "CRITICAL:root:The Run ID can't be higher than 9999. Exiting..."
-            ],
+            ["CRITICAL:root:The Run ID can't be higher than 9999. Exiting..."],
         )
         shutil.rmtree(output_directory)
         self.assertFalse(os.path.isdir(output_directory))
@@ -1078,6 +1077,496 @@ class TestInit(unittest.TestCase):
             shutil.move(input_file, input_directory)
 
         for run_id in (123, 124, 125):
+            # Confirm the output files have been written and their content
+            metadata_file_name = "%s/SSE0%dE.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(metadata_file_name))
+            with open(metadata_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "SSEAAA                           "
+                    "9999999900000000H00000300%dV1.11                          "
+                    "                                                           "
+                    "                                                 " % run_id
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the metadata output file
+            os.remove(metadata_file_name)
+            self.assertFalse(os.path.isfile(metadata_file_name))
+
+            detailed_file_name = "%s/SSE0%dD.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(detailed_file_name))
+            with open(detailed_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "0004000133034205413540000100202007312006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034005407940000157202003051022"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034105409340022139202012252006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           "
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the detailed output file
+            os.remove(detailed_file_name)
+            self.assertFalse(os.path.isfile(detailed_file_name))
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_no_run_id_nor_run_id_file(self):
+        """
+        Test the init code without --run-id nor --run-id-file
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        with self.assertRaises(SystemExit) as cm1, self.assertLogs(
+            level="CRITICAL"
+        ) as cm2:
+            try:
+                target.init()
+            except localeError:
+                # On Windows, the French locale is just called "fr"
+                target.sys.argv[-1] = "fr"
+                target.init()
+        self.assertEqual(cm1.exception.code, 224)
+        self.assertEqual(
+            cm2.output,
+            [
+                "CRITICAL:root:Either the `--run-id` or the `--run-id-file` arguments "
+                "must be specified. Exiting..."
+            ],
+        )
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_new_run_id_file_no_run_id(self):
+        """
+        Test the init code with multiple files, nonexisting --run-id-file and without
+        --run-id. Run ID will start at 0
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        run_id_file = "%s/run-id.txt" % output_directory
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id-file",
+            run_id_file,
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        try:
+            target.init()
+        except localeError:
+            # On Windows, the French locale is just called "fr"
+            target.sys.argv[-1] = "fr"
+            target.init()
+
+        self.assertTrue(os.path.isdir(output_directory))
+        self.assertTrue(os.path.isfile(run_id_file))
+        # The Run ID file and the 6 generated output files
+        self.assertEqual(num_files_in_directory(output_directory), 1 + 6)
+        with open(run_id_file) as f:
+            self.assertEqual("3", f.read())
+
+        # --run-id not defined, so the Run ID starts at 0
+        for run_id in (0, 1, 2):
+            # Confirm the output files have been written and their content
+            metadata_file_name = "%s/SSE000%dE.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(metadata_file_name))
+            with open(metadata_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "SSEAAA                           "
+                    "9999999900000000H0000030000%dV1.11                          "
+                    "                                                           "
+                    "                                                 " % run_id
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the metadata output file
+            os.remove(metadata_file_name)
+            self.assertFalse(os.path.isfile(metadata_file_name))
+
+            detailed_file_name = "%s/SSE000%dD.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(detailed_file_name))
+            with open(detailed_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "0004000133034205413540000100202007312006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034005407940000157202003051022"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034105409340022139202012252006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           "
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the detailed output file
+            os.remove(detailed_file_name)
+            self.assertFalse(os.path.isfile(detailed_file_name))
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_new_run_id_file_and_run_id(self):
+        """
+        Test the init code with multiple files, nonexisting --run-id-file and --run-id
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        run_id_file = "%s/run-id.txt" % output_directory
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id",
+            "123",
+            "--run-id-file",
+            run_id_file,
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        try:
+            target.init()
+        except localeError:
+            # On Windows, the French locale is just called "fr"
+            target.sys.argv[-1] = "fr"
+            target.init()
+
+        self.assertTrue(os.path.isdir(output_directory))
+        self.assertTrue(os.path.isfile(run_id_file))
+        # The Run ID file and the 6 generated output files
+        self.assertEqual(num_files_in_directory(output_directory), 1 + 6)
+        with open(run_id_file) as f:
+            self.assertEqual("126", f.read())
+
+        for run_id in (123, 124, 125):
+            # Confirm the output files have been written and their content
+            metadata_file_name = "%s/SSE0%dE.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(metadata_file_name))
+            with open(metadata_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "SSEAAA                           "
+                    "9999999900000000H00000300%dV1.11                          "
+                    "                                                           "
+                    "                                                 " % run_id
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the metadata output file
+            os.remove(metadata_file_name)
+            self.assertFalse(os.path.isfile(metadata_file_name))
+
+            detailed_file_name = "%s/SSE0%dD.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(detailed_file_name))
+            with open(detailed_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "0004000133034205413540000100202007312006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034005407940000157202003051022"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034105409340022139202012252006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           "
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the detailed output file
+            os.remove(detailed_file_name)
+            self.assertFalse(os.path.isfile(detailed_file_name))
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_invalid_content_existing_run_id_file(self):
+        """
+        Test the init code with existing --run-id-file but with invalid content
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        run_id_file = "%s/run-id.txt" % output_directory
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        pathlib.Path(output_directory).mkdir(parents=True, exist_ok=True)
+        with open(run_id_file, "w") as ofile:
+            ofile.write("INVALID")
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id-file",
+            run_id_file,
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        with self.assertRaises(SystemExit) as cm1, self.assertLogs(
+            level="CRITICAL"
+        ) as cm2:
+            try:
+                target.init()
+            except localeError:
+                # On Windows, the French locale is just called "fr"
+                target.sys.argv[-1] = "fr"
+                target.init()
+        self.assertEqual(cm1.exception.code, 225)
+        self.assertEqual(
+            cm2.output,
+            [
+                "CRITICAL:root:The value stored in the file passed in the "
+                "`--run-id-file` argument must be numeric. Exiting..."
+            ],
+        )
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_existing_run_id_file_no_run_id(self):
+        """
+        Test the init code with multiple files, existing --run-id-file and without
+        --run-id
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        run_id_file = "%s/run-id.txt" % output_directory
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        pathlib.Path(output_directory).mkdir(parents=True, exist_ok=True)
+        with open(run_id_file, "w") as ofile:
+            ofile.write("456")
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id-file",
+            run_id_file,
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        try:
+            target.init()
+        except localeError:
+            # On Windows, the French locale is just called "fr"
+            target.sys.argv[-1] = "fr"
+            target.init()
+
+        self.assertTrue(os.path.isdir(output_directory))
+        self.assertTrue(os.path.isfile(run_id_file))
+        # The Run ID file and the 6 generated output files
+        self.assertEqual(num_files_in_directory(output_directory), 1 + 6)
+        with open(run_id_file) as f:
+            self.assertEqual("459", f.read())
+
+        for run_id in (456, 457, 458):
+            # Confirm the output files have been written and their content
+            metadata_file_name = "%s/SSE0%dE.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(metadata_file_name))
+            with open(metadata_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "SSEAAA                           "
+                    "9999999900000000H00000300%dV1.11                          "
+                    "                                                           "
+                    "                                                 " % run_id
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the metadata output file
+            os.remove(metadata_file_name)
+            self.assertFalse(os.path.isfile(metadata_file_name))
+
+            detailed_file_name = "%s/SSE0%dD.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(detailed_file_name))
+            with open(detailed_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "0004000133034205413540000100202007312006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034005407940000157202003051022"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034105409340022139202012252006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           "
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the detailed output file
+            os.remove(detailed_file_name)
+            self.assertFalse(os.path.isfile(detailed_file_name))
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
+    def test_init_existing_run_id_file_and_run_id(self):
+        """
+        Test the init code with multiple files, existing --run-id-file and --run-id
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        run_id_file = "%s/run-id.txt" % output_directory
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        pathlib.Path(output_directory).mkdir(parents=True, exist_ok=True)
+        with open(run_id_file, "w") as ofile:
+            # This value will not be respected since passing an explicit --run-id
+            ofile.write("456")
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id",
+            "654",
+            "--run-id-file",
+            run_id_file,
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        try:
+            target.init()
+        except localeError:
+            # On Windows, the French locale is just called "fr"
+            target.sys.argv[-1] = "fr"
+            target.init()
+
+        self.assertTrue(os.path.isdir(output_directory))
+        self.assertTrue(os.path.isfile(run_id_file))
+        # The Run ID file and the 6 generated output files
+        self.assertEqual(num_files_in_directory(output_directory), 1 + 6)
+        with open(run_id_file) as f:
+            # The value passed in --run-id is respected, not the value that was
+            # already present in the --run-id-file
+            self.assertEqual("657", f.read())
+
+        for run_id in (654, 655, 656):
             # Confirm the output files have been written and their content
             metadata_file_name = "%s/SSE0%dE.txt" % (output_directory, run_id)
             self.assertTrue(os.path.isfile(metadata_file_name))